net, dns: Move most of the dns subpackage under network
The existing `util/net/dns` subpackage is left with the sanitizer
helper which requires knowledge of the VMI structure.
