+++ pkg/virt-launcher/virtwrap/converter/BUILD.bazel
@@ -21,8 +21,8 @@ go_library(
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
         "//pkg/host-disk:go_default_library",
         "//pkg/ignition:go_default_library",
+        "//pkg/network/dns:go_default_library",
         "//pkg/util:go_default_library",
-        "//pkg/util/net/dns:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
