+++ tests/BUILD.bazel
@@ -178,6 +178,7 @@ go_test(
         "//pkg/hooks:go_default_library",
         "//pkg/hooks/v1alpha1:go_default_library",
         "//pkg/hooks/v1alpha2:go_default_library",
+        "//pkg/network/dns:go_default_library",
         "//pkg/testutils:go_default_library",
         "//pkg/util/cluster:go_default_library",
         "//pkg/util/hardware:go_default_library",
